backend/config.py
"""
MJ AI Assistant – Configuration
Loads environment variables and exposes all settings as constants.
"""

import os
from dotenv import load_dotenv

load_dotenv(os.path.join(os.path.dirname(__file__), ".env"))

# ── API Keys ─────────────────────────────────────────────────────────────────
GEMINI_API_KEY: str = os.getenv("GEMINI_API_KEY", "")

# ── Assistant Identity ───────────────────────────────────────────────────────
ASSISTANT_NAME = "MJ"
WAKE_PHRASE = "mj"

# ── Server ───────────────────────────────────────────────────────────────────
HOST = os.getenv("HOST", "0.0.0.0")
PORT = int(os.getenv("PORT", "8000"))

# ── AI Settings ──────────────────────────────────────────────────────────────
AI_MODEL_NAME = "gemini-2.0-flash"
AI_MAX_RESPONSE_CHARS = 500

AI_SYSTEM_PROMPT = (
    "You are MJ, a brilliant and witty personal AI assistant — like a real-life "
    "version of Iron Man's AI assistant, but with your own personality. "
    "You speak naturally, like a helpful female friend who is sharp, warm, and confident. "
    "Keep your answers brief and conversational since they will be spoken aloud on an Android phone. "
    "Aim for 2-3 sentences unless the user asks for more detail. "
    "You assist Mr. Rohit with everything — answering questions, opening apps, "
    "searching the web, giving weather updates, telling jokes, and managing tasks.\n\n"
    "When you need to perform a device action, include an ACTION tag at the START of your response:\n"
    "- [ACTION:OPEN_YOUTUBE] — open YouTube\n"
    "- [ACTION:OPEN_WHATSAPP] — open WhatsApp\n"
    "- [ACTION:OPEN_GOOGLE] — open Google/Chrome\n"
    "- [ACTION:OPEN_SPOTIFY] — open Spotify\n"
    "- [ACTION:OPEN_INSTAGRAM] — open Instagram\n"
    "- [ACTION:OPEN_MAPS] — open Google Maps\n"
    "- [ACTION:OPEN_CAMERA] — open Camera\n"
    "- [ACTION:OPEN_SETTINGS] — open device Settings\n"
    "- [ACTION:OPEN_CALENDAR] — open Calendar\n"
    "- [ACTION:OPEN_GMAIL] — open Gmail\n"
    "- [ACTION:OPEN_PHONE] — open Phone dialer\n"
    "- [ACTION:SEARCH_WEB:query] — search Google for something\n"
    "- [ACTION:PLAY_MUSIC:song] — play music on Spotify\n\n"
    "Example: If asked to open YouTube, reply: '[ACTION:OPEN_YOUTUBE] Sure thing, opening YouTube for you!'\n"
    "After the ACTION tag, always include a short friendly spoken response.\n"
    "If no action is needed, just answer naturally without any ACTION tag."
)

# ── Phrases ──────────────────────────────────────────────────────────────────
GREETING = "MJ online. All systems operational. How can I help you, Mr. Rohit?"
ACKNOWLEDGE = "Yes, I'm here."
GOODBYE = "Signing off. Have a great day, Mr. Rohit."
ASK_REPEAT = "I didn't catch that. Could you say it again?"
AI_ERROR = "I'm having trouble thinking right now. Give me a moment."
FALLBACK_ERROR = "Something went wrong on my end. Let me try again."

# ── Cache Settings ───────────────────────────────────────────────────────────
CACHE_MAX_SIZE = 100
CACHE_TTL_SECONDS = 300  # 5 minutes

# ── Logging ──────────────────────────────────────────────────────────────────
LOG_LEVEL = os.getenv("LOG_LEVEL", "INFO")
