backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
google-generativeai>=0.5.0
python-dotenv>=1.0.1
pydantic>=2.0
aiofiles>=23.0
